+++ packages/core/components/pull-to-refresh/pull-to-refresh-class.js
@@ -128,7 +128,19 @@ class PullToRefresh extends Framework7Class {
 
       if (!isMoved) {
         $el.removeClass('ptr-transitioning');
-        if (scrollTop > $el[0].offsetHeight) {
+        let targetIsEl;
+        let targetIsScrollable;
+        $(e.target).parents().each((index, targetEl) => {
+          if (targetEl === el) {
+            targetIsEl = true;
+          }
+          if (targetIsEl) return;
+          if (targetEl.scrollHeight > targetEl.offsetHeight) {
+            targetIsScrollable = true;
+          }
+        });
+
+        if (targetIsScrollable || scrollTop > $el[0].offsetHeight) {
           isTouched = false;
           return;
         }