+++ packages/react/components/input.js
@@ -13,6 +13,12 @@ class F7Input extends React.Component {
     super(props, context);
     this.__reactRefs = {};
 
+    this.state = (() => {
+      return {
+        inputFocused: false
+      };
+    })();
+
     (() => {
       const self = this;
       self.onFocusBound = self.onFocus.bind(self);
@@ -48,10 +54,16 @@ class F7Input extends React.Component {
 
   onFocus(event) {
     this.dispatchEvent('focus', event);
+    this.setState({
+      inputFocused: true
+    });
   }
 
   onBlur(event) {
     this.dispatchEvent('blur', event);
+    this.setState({
+      inputFocused: false
+    });
   }
 
   onChange(event) {
@@ -109,7 +121,13 @@ class F7Input extends React.Component {
       const InputTag = tag;
       const needsValue = type !== 'file';
       const needsType = tag === 'input';
-      const inputClassName = Utils.classNames(type === 'textarea' && resizable && 'resizable', !wrap && className, (noFormStoreData || noStoreData || ignoreStoreData) && 'no-store-data', errorMessage && errorMessageForce && 'input-invalid');
+      const inputClassName = Utils.classNames(!wrap && className, {
+        resizable: type === 'textarea' && resizable,
+        'no-store-data': noFormStoreData || noStoreData || ignoreStoreData,
+        'input-invalid': errorMessage && errorMessageForce,
+        'input-with-value': typeof value === 'undefined' ? defaultValue || defaultValue === 0 : value || value === 0,
+        'input-focused': self.state.inputFocused
+      });
       let input;
       {
         input = React.createElement(InputTag, {