+++ scripts/release.sh
@@ -19,22 +19,26 @@ if [[ $REPLY =~ ^[Yy]$ ]]; then
 
   # remove old builds
   rm -rf ./packages/core/components
+  rm -rf ./packages/core/lazy-components
   rm -rf ./packages/core/css
   rm -rf ./packages/core/js
   rm -rf ./packages/core/less
   rm -rf ./packages/core/modules
   rm -rf ./packages/core/utils
   rm -rf ./packages/core/*.js
+  rm -rf ./packages/core/*.ts
   rm -rf ./packages/core/*.less
   rm -rf ./packages/react/components
   rm -rf ./packages/react/runtime-helpers
   rm -rf ./packages/react/utils
   rm -rf ./packages/react/*.js
+  rm -rf ./packages/react/*.ts
   rm -rf ./packages/react/*.map
   rm -rf ./packages/vue/components
   rm -rf ./packages/vue/runtime-helpers
   rm -rf ./packages/vue/utils
   rm -rf ./packages/vue/*.js
+  rm -rf ./packages/vue/*.ts
   rm -rf ./packages/vue/*.map
 
   # build
@@ -62,6 +66,9 @@ if [[ $REPLY =~ ^[Yy]$ ]]; then
       npm publish
     )
   )
+
+  # Build Production Kitchen Sink
+  npm run ks:prod
 
   # commit
   git add -A